Regenerate add_sources.tcl and resume .do script if there is an error
Resuming the QuestaSim simulation script in case there is an error
is a great improvement to the flow. Usually the error is due
to a signal not existing when an 'add wave' command is given.
By resuming the script, we can add signals that might not be available
for all simulations.

Regenerating add_sources.tcl, incase the source list (.sv files for example)
is changed. Otherwise, the new files won't be compiled and simulated.