Rename dcache, reorganize, finish data cache (testing in sim)
Passes verilator linting (warnings waivered).

Fixed compilation errors with Xilinx, testing in Questasim

There is a store and then a load request, and even though we grant the
load request, we still immediately for some reason. I don't know why we
stall immediately. It might be because we need the data immediately.
Check original sim.

Original sim shows a stall immediately after request is issued.
So I guess CPU needs data before starting. So we need to debug
the cache and see why its not reading from main memory
or if the states are correct.