Comment out everything thats unnecessary to test memory interface
Doing this to debug the bypassed cache interface. We should be able
to complete the program (at least the start of it) by forwarding all
memory operations to the memory request output. Once this works, we
can start adding in the other pieces (tag store, data store, tag
comparision, etc.)