cpu <--> memory interface seems to work, but MNIST UART does not output?
Fix immediately serving kill req with valid signal

...and that seems to do the trick!!! Simulation is no longer stalling,
seems like some actual work is being done. I didn't realize that the
kill_req was hanging up Ariane, but it made sense once I saw the code that
waited for a rvalid signal, and also when I further analyzed the simulation
noticing repetitive memory accesses.

However, I just ran a simulation for 180 ms and although the end of the simulation
looks correct (we stop fetching requests and the number of loads and stores
doesn't change, we didn't get UART output.

I have a suspicion it's because I applied a "breakpoint" in questasim by clicking
break simulation to increase the length of the simulation. I'm not certain it was that
though... I"m currently redoing the simulation.

In between that time, I'm going to create a new branch and start to slowly uncomment
the cache changes. If the UART application works, I know that the input and output
of the cache system is functioning... If it doesn't, well... lets hope it is.