Focus on bypassing cache and thus only keep code related to that
Existing code is extremely simplified and includes almost 0 signals related
to cache handling (cache hit, etc.). The idea was to simplify the code
so that I don't get bogged down with possible errors related to enabling
the cache, etc.

Also cleaned up dcache_pkg, to only include functions that concern CPU
to main memory address interfacing, and for selecting appropiate words
from the returning cache block.

Finally updated waveform viewing.

Tested simulation as of 04/29/22 around 11h00. Able to complete the simulation
and get past the debug self test. However, the UART message never shows, which leads me to
believe that the simulation did not complete correctly (i.e. the data was not correctly
loaded or stored).

With that said, looks like the simulation does "end" so to speak, at around
175 ms, which indicates to me that we reached the end of the MNIST application. The end of this
simulation is similiar to the openhwgroup branch end of simulation, (the golden sim dataset).

One thing to check is that the correct addresses are being written and read from, by using
QuestaSim's block memory viewer, and comparing contents of the block RAM. i.e. manually checking
the addresses accessed, to assure that they match with what addresses and data are being put out
during the simulation.

Ok trying the above memory thing seems complicated, especially because the memory addresses are not 1:1.

One thing I just checked, was the activity on the JTAG bus. Seems at the end of the golden dataset,
there is a bunch of TX occuring. However, on the bypass cache, there is a little TX but not
nearly as much. I'm thinking that the TX addresses correspond to I/O space, and I might have
to do some special formatting of those addresses or data sizes to output to the UART which
is in I/O space (thats my guess anyway).