Add correct CPU word selection from returned cache block - sim works
The final piece of the puzzle that was blocking the sim from working was
that we were selecting the incorrect data after the loads from main memory.
I noticed something was fishy when I added the JTAG signals and the beginning part
of the simulation had a bunch of TXs occuring, where as with the golden sim
there were 0 transfers in the beginning.

I realized that the way I was indexing into the cache block returned via the
mem_data_o.data field was rather incorrect, especially for load byte instructions.
With that said, I haven't yet analyzed the code I took from wt_dcache_mem, to
figure out exactly how the indexing works. All I know is it works now.

Seems that the full program takes about 185 ms, so sim time is set to that.

Post impl sim results show about 45.6 mJ energy consumption. What's interesting
to remark though is the fairly reduced power consumption (0.264 W) compared to
the baseline (0.301 W). If we can compensate the performance loss (3.7x
increase in number of clock cycles) I think we can get some pretty good energy
performance.