Work in progress - UART messages spitting out gibberish
So went ahead and double checked simulation by adding in a force bypass
signal. When I fully bypass the cache, everything works and I get the expected
UART message. So I at least know that the problem is somewhere within the cache
and the new code I put.

I had spotted a pretty important error, which was related to writeback. That is,
I wasn't using the correct data to writeback to main memory, because I hadn't
had a clock cycle for the data to come out of the data store.

UART messages are still spitting out garbage though around 90 ms or so. After
2 UART messages I get errors from some asserts within the wt_cache_subsystem module,
mentioning incorrect memory transactions/data/instructions, etc. So something is
clearly wrong somewhere.

I am going to go through the code line by line and make sure it makes sense before
looking at the simulation again. It's tough to look at the simulation when it
comes to a writeback cache, as the data is not immediately written back to main memory
and it's difficult to keep track of the previous cache accesses.

Perhaps some actual assertions or UVM testbenching ? Although I'm not sure if
that's going to have too steep a learning curve and not be worth it.