Perform full simulation and get results for basic riscmakers cache
So I'm not 100% sure that the cache is working without fault, although
it does seem to be, because the UART messages are being output
correctly.

So currently the cache is underperforming the write through cache,
but not by a whole lot. And there are plenty of oppourtunities
to improve the performance.