Change cache parameter defines and behav test sim with a larger cache
Instead of having the cache parameters separated into the dcache_pkg I
decided to simply change the parameters within ariane_pkg (because
they are used throughout the design and not just in the cache part)
and then use the wt_cache_pkg for the calculated parameters.

Eventually I will use my own risckmakers_pkg for the calculated
parameters, but because we're using the wt_axi_adapter, I thought
it best to keep using the wt_cache_pkg.

Just ran behav sim. The number of clock cycles went down to
2349428, and the number of cache misses to 4955. Note that the
cache size was increased from 4 KiB to 32 KiB. Need to run
post impl to see the power consumption change.