Convert bit enable tag store to byte enable tag store - sim working
Realized after getting the riscmakers dcache working that the number of BRAMs
instanciated for the tag store didn't make sense. It seemed like way too many, and
after doing the calculation for the maximum amount needed, the actual number was like
10x that.

I figured it could be for 2 reasons. Either it was how I coded up the single port
BRAM (perhaps I was incorrectly inferring it?) or it was because I was using a bit
enable signal, instead of the byte enable that seems to be the only option supported
by the Xilinx block RAM. Spoiler alert: it was the latter. I shouldn't have assumed
the tool was going to synthesize logic to allow for bit enabling. Well, I mean it
did do that, I think it just synthesized like 8 more block rams to allow for it.

So my solution was simply to keep the main structure the same (tag store structure,
setting the tag store bits through this structure, etc.) but then add a second
structure that was byte aligned and use that to hook into the instanciated Xilinx BRAM.

Found a couple of improvements/changes along the way. For example, removed within the
Xilinx BRAM defining the tag store types (not necessary, and it kind of defeats the purpose
of having a BRAM that is a bit generic). Also, realized I don't need to have all these
bit enable bits for the tag store write enable tag. I could just have a single bit, and then
that bit is used to enable a write to all tag store bytes.

Behavioral sim works, UART message is correctly transmitted. Post impl shows 2 mW
improvement.

Ran OOC FPGA synthesis and it looks like the number of BRAMs for tag store is now
working (also coherent with data store). 8x36 BRAMs for the data store, which makes
sense since we have a 32 KiB cache, and then 2x36 BRAMs and 1x18 BRAM. This also makes
sense because with a tag size of 17 bits, when it is byte aligned along with the
dirty and valid bit, we get 40 bits. 40 bits times the depth (2048, or the number
of cache sets) comes out to 81920 bits, and thus requires the instanciated amount
of block RAM.

Initially ran into some trouble with this, because I had assigned the valid bit twice
instead of the valid and dirty bit. But I fixed it. The error was subtle though, because
simulation seemed to run fine, the jtag TX looked normal, core resumes, etc., but then
but then I get a UART message: illegal instruction exception handler entered