Make the cache block size configurable, and serve consecutive requests
Got full simulations results for riscmakers dcache 32 KiB, 256 bit cache block.
Calculated roughly 13.67 mJ consumption, which when compared to baseline
gives a 4.54% energy improvement.

Successively able to choose the block size at will. I tried (without
commiting results) a cache block size of 128, 256, and 512. All work
without problem. The number of cache misses decreases by roughly
2 each time. There is a rather remarkable change between
128 and 258 bits, not sure about 258 to 512. Also not
sure if we can increase beyond 512... I don't think so. At any rate, increasing
the cache block size did not change the power consumption, but did improve the
application performance. Another full sim with a cache block size of 512 will be carried out.

Able to serve back to back loads. This turned out to be huge
for the performance, because in the simulation there are great deal
of back to back loads. Being able to serve cache load hits like that
reduced the number of clock cycles by like 300,000! I ran a quick
behav simulation and got 2092382 clock cycles for the program
execution. So we're below the number of clock cycles as per the golden value.

However... I didn't realize before, but we actually have a slack timing violation.
It's roughly 1-2  ns, and it's due to the address bits coming from the load/store unit
going to the tag store. Because the tag store is sensitive to data on the negative edge,
there is a setup violation (as the address bits have a delay reaching the tag store).
1 solution to overcome this is simply reduce the clock cycle frequency (increase the period)
to account for this 1 ns violation. However we need to be careful to still respect
the minimal performance.

Reducing the clock frequency though might have a large positive impact on the power consumption.
Specifically, the dynamic power. Currently it's around 182 mW, and takes up the majority of
the consumed energy. So reducing the clock frequency might actually have a second benefit,
in addition to overcoming the timing violation.

Another solution might be to reduce the logic path from the load and store unit. I'm especially
suspicious of the remaining logic used for physical address translation, that are clearly not
necessary if the MMU is not enabled. Either that, or perhaps changing the size of the
depth commits, of the number of load/store pipeline registers, etc. I noticed that there
are 0 store pipeline registers, and a quick glance at the cva6_sim.timing.rpt shows
quite a few logic delays through the store unit. I'll have to investigate this.

Finally, note that I accidentally deleted the behav/ simulation results in one of the previous
commits. I readded it for the 256 bit cache block sim.