Refactor data cache, energy improvement is 5.87%
Now the data cache can serve requests more easily. The stores are also
modified to be almost identical (minus the valid bit being set to 0
for the tag store). Finally, cleaned up various things in the source
code (old comments, etc.).

Was a bit surprised to see a power decrease of 1 mW, especially
since the power was not decreased for the cache system, but for
other components (peripherals, etc.).