WIP - convert dcache into an icache (behav sim works with bypassed cache)
Convert the riscmakers dcache into a replacement for the cva6_icache

Currently only bypassing cache (to assure that the communication and flow of data between input
and output occurs correctly).

Full simulation takes roughly 527 ms to output data from the UART. Results are all
correct and number of clock cycles is 23095265 (extra cycles due to the icache being bypassed).

Next step is to enable icache and test sim.