WIP - icache working (behav sim works)
Takes about 90 ms for the simulation to complete. Results
are all correct with cache enabled, and there are 3468705 clock cycles,
and 1002 I$ misses.

Because there are roughly the same number of I$ as there are in the
original I$, I'm thinking that the extra latency is due to the
cache not handling back to back cache hits, or handling a new request
as soon as possible. I'm going to test this theory.