Run behav sim after adding serve_new_request()
So sim completes after roughly 60 ms, results are correct, but I did get a error output
from the sim terminal before the core resumed:

** Error: Access to register 07b1 failed with error 3
   Time: 67156 ns  Scope: jtag_pkg.debug_mode_if_t.wait_command File: ../../../../../tb/jtag_pkg.sv Line: 777
[TB]             68908000 - Resuming the CORE

I'm thinking it is because I was serving a new request as soon as there was a kill request?
Looking at the git diff, I don't see any huge changes I made to the I$. Since there wasn't this
error before, I'm guessing thats the reason.

For now I'm going to keep testing, and run full post impl just to see if the energy is reduced.
If so, I'll continue with actually testing the cache on the target device, to see if the error
actually matters.