Run post impl for riscmakers icache - energy improvement 7.12%
Still need to run actual execution on Zybo to confirm if the changes
are allowed, but currently sitting at 7.12% energy improvement.

Do note however that a combinatorial loop was found:
"There is 1 combinational loop in the design" as mentioned
in cva6_sim.check_timing.rpt. In log file, this is located here:
[/home/justin/riscmakers/cva6-softcore-contest/include/ariane_pkg.sv:99]
and thus I think the issue is either:

is_inside_nonidempotent_regions();
or
is_inside_cacheable_regions();

Perhaps the addr_ni or mem_data_o.nc determines the logic of req_address_d ? I don't
think addr_ni is an issue because we are using the req_address_q output, and thus
it should be stable after going through a register.

Although log does say: set_disable_timing req_address_q[13]_i_4,
so perhaps it is an issue with req_address_q...

Yes it seems as though the issue is with req_address_q. Although it
seems to be due to icache_adapter[nc]... and thus not addr_ni but
the nc bit.

I'm thinking it might be due to the en_i bit... I see in cva6_icache
that this bit was registered. Although it is talking about req_address_q
so it might be that... although in cva6_icache this doesn't seem to be using the
registered value.

Going to change icache parameters to see how this affects energy