Fix error occuring from tb/jtag and error preventing OpenOCD from working - emulation works
I'll be honest, not 100% sure of why the error was coming about, but found the location of it.
It was in the WAIT_NON_SPECULATIVE_FLAG portion of the code. Serving a new request in this situation
(I guess when there there is a non-cacheable request and it is speculative) resulted in error output.
So i just changed this to not serve a new request immediately and voila. No error from behav sim.

Also ran the application on the FPGA target and successively able to run the MNIST application. Due
to this error, I'm assuming, OpenOCD could not run.

Also tried rerunning the application and that worked to. A couple less cache misses but otherwise pretty
much the same results obtained. So it can be rerun.