Decrease riscmakers I$ size to 8 KiB - energy improvement of 9.71%
Surprisingly, decreasing the instruction cache size by half did not change
the number of clock cycles needed to execute the MNIST application. I'm
thinking that this is because the missed instructions that resulted in
extra clock cycles (which seems to be obligatory) are simply not
accounted for in summary of the MNIST execution statistics. Not sure.

Anyway, this saved a mW of power. I'm going to attempt a 4 KiB size
behav sim, just to see if this changes the number of clock cycles or not.