Reduce I$ size to 2 KiB - full sim
Wow, again only a dozen or so increase in the number of clock cycles
for the execution of the MNIST application. With a 2 KiB cache,
for the data store that should only take a single 18K BRAM (16384 bits
needed in total). For the tag store, I calculated only 4096 bits needed,
and thus again only a single 18K BRAM.

With that, I don't think it makes any more sense to decrease the size further, as
the 18K BRAM will still be instantiated. Perhaps that'd be interesting to see
with a 1 KiB I$.

I do notice however that the total number of loads for the cache is decreasing
as the size of the I$ decreases. This is a bit surprising for me, but I guess
it has something to do with the size of the index and tags bits. I'm guessing...

Post impl sim is surprising - 2 36K BRAMs are being instantiated for a 128 x 128
block RAM. I would of thought that this could fit with a single 18K BRAM. Perhaps
the tool is trying to use a low power optizmiation instead of smallest area?

I'm going to run the synthesis with the lower power optizimations (for BRAM only)
just to see.